move interfaces into abstractions folder
