+++ LLama/ChatSession.cs
@@ -1,4 +1,5 @@
-﻿using LLama.Common;
+﻿using LLama.Abstractions;
+using LLama.Common;
 using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
