+++ LLama/LLamaExecutorBase.cs
@@ -1,4 +1,5 @@
-﻿using LLama.Common;
+﻿using LLama.Abstractions;
+using LLama.Common;
 using LLama.Exceptions;
 using LLama.Native;
 using System;
